requirements.txt
PyMuPDF
freeGPT
reportlab
tgcrypto
git+https://github.com/Xcilik/smallgram.git
pytz
SpeechRecognition
pydub
py-tgcalls==1.0.9
ntgcalls==1.1.3
git+https://github.com/yt-dlp/yt-dlp.git
google_currency
speedtest
youtube-search-python
motor
dnspython
GitPython
pykeyboard
bs4
Pillow==9.5.0
pymediainfo
opencv-python-headless
requests
rich
unidecode
attrify
pyromod
html2text
removebg
aiofiles
aiohttp
gpytranslate
gtts
wget
telegraph
pytimeparse
moviepy
pymongo==4.6.3
